+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Updates
 
-- Add the webpack `v5` compatibility.
+- Add the webpack `v5` compatibility and the backward compatibility for the `v4`.
 - Use the `processAssets` compilation hook to create templates and the `chunks-manifest.json` file.
 - The plugin requires at least Node.js `10.13.0` (LTS). The `engines.node` is updated in the `package.json`.
 - Update the nodeJS version in the Github Action